+++ plugin/seq/msh3.cpp
@@ -2002,32 +2002,41 @@ class listMeshS {
   }
 };
 
-class listMeshL {
+template<class MeshT>
+class listMeshT {
 public:
-    list< const MeshL * > *lth;
-    void init( ) { lth = new list< const MeshL * >; }
+    typedef MeshT Mesh;
+    typedef const MeshT *pmeshT;
+    typedef list< const MeshT * > List;
+    list< const MeshT * > *lth;
+    void init( ) { lth = new list< const MeshT * >; }
     
     void destroy( ) { delete lth; }
     
-    listMeshL(Stack s, const MeshL *th) : lth(Add2StackOfPtr2Free(s, new list< const MeshL * >)) {
+    listMeshT(Stack s, const MeshT *th) : lth(Add2StackOfPtr2Free(s, new list< const MeshT * >)) {
         lth->push_back(th);
     }
     
-    listMeshL(Stack s, const MeshL *tha, const MeshL *thb)
-    : lth(Add2StackOfPtr2Free(s, new list< const MeshL * >)) {
+    listMeshT(Stack s, const MeshT *tha, const MeshT *thb)
+    : lth(Add2StackOfPtr2Free(s, new list< const MeshT * >)) {
         lth->push_back(tha);
         lth->push_back(thb);
     }
     
-    listMeshL(Stack s, const listMeshL &l, const MeshL *th)
-    : lth(Add2StackOfPtr2Free(s, new list< const MeshL * >(*l.lth))) {
+    listMeshT(Stack s, const listMeshT &l, const MeshT *th)
+    : lth(Add2StackOfPtr2Free(s, new list< const MeshT * >(*l.lth))) {
         lth->push_back(th);
+    }
+    listMeshT(Stack s, KN< pmeshT > *const &tab)
+    : lth(Add2StackOfPtr2Free(s, new list< const MeshT * >)) {
+        for(int i=0; i<tab->N();++i)
+          lth->push_back((*tab)[i]);
     }
 };
 
 
 
-MeshS *GluMeshS(listMeshS const &lst) {
+MeshS *GluMesh(listMeshT<MeshS> const &lst) {
     typedef typename MeshS::Element T;
     typedef typename MeshS::BorderElement B;
     typedef typename MeshS::Vertex V;
@@ -2177,7 +2186,7 @@ MeshS *GluMeshS(listMeshS const &lst) {
     
 
 
-MeshL *GluMeshL(listMeshL const &lst) {
+MeshL *GluMesh(listMeshT<MeshL> const &lst) {
     typedef typename MeshL::Element T;
     typedef typename MeshL::BorderElement B;
     typedef typename MeshL::Vertex V;
@@ -2338,7 +2347,7 @@ template< bool INIT, class RR, class AA = RR, class BB = AA >
 struct Op3_setmeshS : public binary_function< AA, BB, RR > {
   static RR f(Stack stack, const AA &a, const BB &b) {
     ffassert(a);
-    const pmeshS p = GluMeshS(b);
+    const pmeshS p = GluMesh(b);
 
     if (!INIT && *a)
       (**a).destroy( );
@@ -2357,7 +2366,7 @@ template< bool INIT, class RR, class AA = RR, class BB = AA >
 struct Op3_setmeshL : public binary_function< AA, BB, RR > {
     static RR f(Stack stack, const AA &a, const BB &b) {
         ffassert(a);
-        const pmeshL p = GluMeshL(b);
+        const pmeshL p = GluMesh(b);
         
         if (!INIT && *a)
             (**a).destroy( );
@@ -7382,6 +7391,40 @@ struct Op_GluMesh3tab : public OneOperator {
 
   Op_GluMesh3tab( ) : OneOperator(atype< const pmesh3 >( ), atype< KN< pmesh3 > * >( )){};
 };
+
+template<class MeshT>
+struct Op_GluMeshTtab : public OneOperator {
+
+  typedef const MeshT *pmeshT;
+  class Op : public E_F0mps {
+   public:
+    //static basicAC_F0::name_and_type name_param[];
+    static const int n_name_param = 0;
+    Expression nargs[1];//  use
+    Expression getmeshtab;
+    long arg(int i, Stack stack, long a) const {
+      return nargs[i] ? GetAny< long >((*nargs[i])(stack)) : a;
+    }
+
+    Op(const basicAC_F0 &args, Expression t) : getmeshtab(t) {
+      args.SetNameParam(n_name_param);
+    }
+
+    AnyType operator( )(Stack s) const {
+        KN< const MeshT * > *tab = GetAny< KN< const MeshT * > * >((*getmeshtab)(s));
+        listMeshT<MeshT> lst(s,tab);
+        MeshT *Tht = GluMesh(lst);
+        Add2StackOfPtr2FreeRC(s, Tht);
+        return Tht;
+      }
+  };
+
+  E_F0 *code(const basicAC_F0 &args) const { return new Op(args, t[0]->CastTo(args[0])); }
+
+    Op_GluMeshTtab( ) : OneOperator(atype< const pmeshT >( ), atype< KN< pmeshT > * >( )){};
+};
+
+
 basicAC_F0::name_and_type Op_GluMesh3tab::Op::name_param[Op_GluMesh3tab::Op::n_name_param] = {
   {"labtodel", &typeid(long)}};
 AnyType Op_GluMesh3tab::Op::operator( )(Stack stack) const {
@@ -7392,6 +7435,8 @@ AnyType Op_GluMesh3tab::Op::operator( )(Stack stack) const {
   Add2StackOfPtr2FreeRC(stack, Tht);
   return Tht;
 }
+
+
 
 // return nbc the number of conex componants
 long BuildBoundaryElementAdj(const MeshS &Th, bool check = 0, KN< long > *pborder = 0) {
@@ -9422,6 +9467,9 @@ class OrientNormal : public OneOperator {
 // <<dynamic_loading>>
 
 static void Load_Init( ) {
+ typedef listMeshT<MeshL> listMeshL;
+ typedef listMeshT<MeshS> listMeshS;
+
   Dcl_Type< listMesh3 >( );
   Dcl_Type< listMeshS >( );
   Dcl_Type< listMeshL >( );
@@ -9449,6 +9497,8 @@ static void Load_Init( ) {
 
   Global.Add("trunc", "(", new Op_trunc_mesh3);
   Global.Add("gluemesh", "(", new Op_GluMesh3tab);
+  Global.Add("gluemesh", "(", new Op_GluMeshTtab<MeshL>);
+  Global.Add("gluemesh", "(", new Op_GluMeshTtab<MeshS>);
   Global.Add("extract", "(",new ExtractMesh< Mesh3, MeshS >);    // take a Mesh3 in arg and return a part of MeshS
 
   // for a mesh3 Th3, if Th3->meshS=NULL, build the meshS associated